+++ llimits.h
@@ -271,7 +271,7 @@ union luai_Cast { double l_d; LUA_INT32 l_p[2]; };
 /* on several machines, coercion from unsigned to double is slow,
    so it may be worth to avoid */
 #define lua_unsigned2number(u)  \
-    (((u) <= (lua_Unsigned)INT_MAX) ? (lua_Number)(int32_t)(u) : (lua_Number)((int32_t)u))
+    (((u) <= (lua_Unsigned)SHRT_MAX) ? (lua_Number)(int16_t)(u) : (lua_Number)(int16_t)u)
 #endif
